feat(DSTEW-1595): - Class full path removal

+++ claims-data/service/src/main/java/uk/gov/justice/laa/dstew/payments/claimsdata/config/ApiProperties.java
@@ -17,5 +17,5 @@ public class ApiProperties {
   private final String url;
   private final String accessToken;
   private final String authHeader;
-  private final long readTimeoutMs;
+  private final int readTimeoutMs;
 }
